add rspec, rewrite spec
