+++ Gemfile
@@ -6,11 +6,12 @@ gem 'benchmark', '0.1.0'
 gem 'byebug', '11.1.3'
 gem 'fasterer', '0.8.3'
 gem 'minitest', '5.14.2'
-#gem 'oj'
-#gem 'rspec-benchmark'
+gem 'pry'
+gem 'rspec'
+gem 'rspec-benchmark'
 gem 'rubocop'
 gem 'rubocop-performance'
 gem 'ruby-prof', '1.4.1'
-#gem 'ruby-progressbar', '~> 1.10', '>= 1.10.1'
+gem 'ruby-progressbar', '~> 1.10', '>= 1.10.1'
 gem 'stackprof', '0.2.15'
